Add getting & setting controller name
